Check for zarr in `images_to_stack` and use da.stack instead of np.stack (#8267)
# References and relevant issues
Closes: https://github.com/napari/napari/issues/8021
see also discussion at:
https://napari.zulipchat.com/#narrow/channel/212875-general/topic/handling.20zarr.2Fdask.2Fetc.20when.20splitting.20or.20merging.20layers/with/532610558

# Description
So thanks to NEP-18 dispatch, `np.stack` will use the `stack` of other
array types automatically, e.g. dask.array. However, zarr doesn't have a
stack method, so then np.stack is used after all.
This PR adds a check to see if the input images have zarrs as image.data
or image.data[0] (to handle multiscale) and if so, use dask.array.stack
instead of np.stack -- with a warning for the user.
This way, zarrs can be merged to a stack with lazy behavior maintained.

---------

Co-authored-by: Lorenzo Gaifas <brisvag@gmail.com>